Mingw-win32 compilation + cleaning

+++ src/main.cpp
@@ -51,7 +51,7 @@ int main()
                 // on attrape les évènements de redimensionnement
                 if (event.type == sf::Event::Resized)
                 {
-                    g_cameraManager.GetCamera()->setSize(event.size.width, event.size.height);
+                    g_cameraManager.GetCamera()->setSize((float)event.size.width, (float)event.size.height);
                 }
                 else if (event.type == sf::Event::Closed)
                 {
@@ -68,7 +68,6 @@ int main()
 
             g_drawManager.Draw();
 
-            //pGameBoard->DbgDisplayGrid(false);
             pGameBoard->DbgDrawCenter();
 
             window.display();
